+++ distributed-core/core-service/src/main/java/com/distributed/transaction/module/account/domain/PayAccountHistoryEntity.java
@@ -37,7 +37,7 @@ public class PayAccountHistoryEntity implements Serializable {
 
     private Date editTime;
 
-    private long version;
+    private Integer version;
 
     private String remark;
 
@@ -64,6 +64,7 @@ public class PayAccountHistoryEntity implements Serializable {
     private String userNo;
 
     private String status;
+
 
     @Id
     @GenericGenerator(name = "system-uuid", strategy = "uuid")
@@ -108,12 +109,12 @@ public class PayAccountHistoryEntity implements Serializable {
     @Basic
     @Column(name = "version")
     @Version
-    public long getVersion() {
+    public Integer getVersion() {
 
         return version;
     }
 
-    public void setVersion(long version) {
+    public void setVersion(Integer version) {
 
         this.version = version;
     }
